fwr: Changed the data type of entry_media_type dimension to support the join to dwh_hourly_events_devices (inc. installation script). Updated installation script to feature 6115 as well.
git-svn-id: svn+ssh://kelev.kaltura.com/usr/local/kalprod/dwh/core/trunk@7044 09e43ab6-4768-4770-81d3-4e3a7c8fdb74

+++ ddl/dw/dw_ENTRY_MEDIA_TYPE.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS `kalturadw`.`dwh_dim_entry_media_type`;
 
 CREATE TABLE `kalturadw`.`dwh_dim_entry_media_type` (
-  `entry_media_type_id` SMALLINT NOT NULL ,
+  `entry_media_type_id` INT(11) NOT NULL ,
   `entry_media_type_name` VARCHAR(50) DEFAULT 'missing value',
    dwh_creation_date TIMESTAMP NOT NULL DEFAULT 0,
    dwh_update_date TIMESTAMP NOT NULL DEFAULT NOW() ON UPDATE NOW(),
